UnavailabilityService tests finished

+++ src/main/java/org/ordina/ordinaForKids/unavailability/UnavailabilityServiceImpl.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.ordina.ordinaForKids.calendarEvent.CalendarEvent;
-import org.ordina.ordinaForKids.calendarEvent.CalendarEventServiceImpl;
+import org.ordina.ordinaForKids.calendarEvent.CalendarEventService;
 import org.ordina.ordinaForKids.validation.UnavailabilityNotFoundException;
 import org.ordina.ordinaForKids.validation.UnavailabilityOverlapsEventException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,18 +20,22 @@ public class UnavailabilityServiceImpl implements UnavailabilityService {
 	UnavailabilityRepository unavailabilityRepository;
 
 	@Autowired
-	CalendarEventServiceImpl calendarEventServiceImpl;
+	CalendarEventService calendarEventService;
 
 	@Override
 	@Transactional
-	public void createUnavailability(Unavailability unavailability) throws UnavailabilityOverlapsEventException {
+	public void createUnavailability(Unavailability unavailability) throws UnavailabilityOverlapsEventException, IllegalArgumentException {
 		checkEventOverlap(unavailability);
-		unavailabilityRepository.save(unavailability);
+		if (isInRepository(unavailability)) {
+			throw new IllegalArgumentException();
+		} else {
+			unavailabilityRepository.save(unavailability);
+		}
 	}
 
 	private void checkEventOverlap(Unavailability unavailability) throws UnavailabilityOverlapsEventException {
 		LocalDate unavailabilityDate = unavailability.getDate();
-		Collection<CalendarEvent> eventsInDatabase = calendarEventServiceImpl.getCalendarEvents();
+		Collection<CalendarEvent> eventsInDatabase = calendarEventService.getCalendarEvents();
 		for (CalendarEvent calendarEvent : eventsInDatabase) {
 			LocalDate calendarEventDate = calendarEvent.getDate();
 			if (calendarEventDate.equals(unavailabilityDate)) {
@@ -40,10 +44,24 @@ public class UnavailabilityServiceImpl implements UnavailabilityService {
 		}
 	}
 
+	private boolean isInRepository(Unavailability unavailability) {
+		long id = unavailability.getId();
+		try {
+			getUnavailability(id);
+			return true;
+		} catch (UnavailabilityNotFoundException e) {
+			return false;
+		}
+	}
+
 	@Override
 	@Transactional
-	public void deleteUnavailability(Unavailability unavailability) {
-		unavailabilityRepository.delete(unavailability);
+	public void deleteUnavailability(Unavailability unavailability) throws UnavailabilityNotFoundException {
+		if (isInRepository(unavailability)) {
+			unavailabilityRepository.delete(unavailability);
+		} else {
+			throw new UnavailabilityNotFoundException(unavailability.getId());
+		}
 	}
 
 	@Override
@@ -52,12 +70,12 @@ public class UnavailabilityServiceImpl implements UnavailabilityService {
 		if (unavailability.isEmpty()) {
 			throw new UnavailabilityNotFoundException(id);
 		}
-		return unavailability.get(); 
+		return unavailability.get();
 	}
 
 	@Override
 	public List<Unavailability> getAllUnavailabilities() throws UnavailabilityNotFoundException {
-		List<Unavailability> unavailabilities = unavailabilityRepository.findAll(); 
+		List<Unavailability> unavailabilities = unavailabilityRepository.findAll();
 		if (unavailabilities.isEmpty()) {
 			throw new UnavailabilityNotFoundException();
 		}
